requirements.txt
Package             Version
------------------- -------
asgiref             3.2.10
Django              3.0.8
django-cors-headers 3.4.0
djangorestframework 3.11.0
Pillow              7.2.0
pip                 20.1.1
pytz                2020.1
setuptools          47.3.1
sqlparse            0.3.1
wheel               0.34.2
